add: enum for win state, game loop, and game logic
Added a tuple with values about the win state of the game, to use as an
enum-like. Added a simple game loop, user input, and checks that the
input the user entered is permitted. Added the game logic, including the
NPC-player generating a selection from "rock," "paper," and "scissors,"
and then comparing the user's selection and the NPC's selection to
choose a win-state; the user is then informed as to whether they won,
lost, or drew.